requirements.txt
future
requests
attrs
